# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "ASP - v2",
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/invitations/1": {
      "get": {
        "summary": "Get Invitation Details",
        "tags": [
          "Users"
        ],
        "operationId": "GetInvitationDetails",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "Accept Invitation",
        "tags": [
          "Users"
        ],
        "operationId": "AcceptInvitation",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AcceptInvitationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/bugs": {
      "post": {
        "summary": "Create Bug",
        "tags": [
          "Bugs"
        ],
        "operationId": "CreateBug",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keyConnection",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateBugRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "Get Bugs",
        "tags": [
          "Bugs"
        ],
        "operationId": "GetBugs",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/bugs/2": {
      "put": {
        "summary": "Update Bug",
        "tags": [
          "Bugs"
        ],
        "operationId": "UpdateBug",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateBugRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/preferences": {
      "post": {
        "summary": "Add/Update Preference",
        "tags": [
          "Notifications"
        ],
        "operationId": "Add/UpdatePreference",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Add~1UpdatePreferenceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "Get Preferences",
        "tags": [
          "Notifications"
        ],
        "operationId": "GetPreferences",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/reports/critical": {
      "get": {
        "summary": "Critical",
        "tags": [
          "Reports"
        ],
        "operationId": "Critical",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "keyConnection",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/costs": {
      "get": {
        "summary": "Get Costs",
        "tags": [
          "CostsExplorer"
        ],
        "operationId": "GetCosts",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "To",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
...
    "AcceptInvitationRequest": {
      "title": "AcceptInvitationRequest",
      "example": {
        "email": "yuyo@mailinator.com",
        "name": "Maias",
        "password": "bobaina"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "name",
        "password"
      ]
    },
...
    "CreateBugRequest": {
      "title": "CreateBugRequest",
      "example": {
        "title": "{{$randomCompanyName}}",
        "description": "{{$randomLoremSentences}}",
        "severity": 1
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "severity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "title",
        "description",
        "severity"
      ]
    },
    "UpdateBugRequest": {
      "title": "UpdateBugRequest",
      "example": {
        "id": 2,
        "title": "New Stark",
        "description": "description",
        "severity": 1,
        "environmentId": 1,
        "systemId": 1,
        "userId": 1,
        "stateId": 1
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "severity": {
          "type": "integer",
          "format": "int32"
        },
        "environmentId": {
          "type": "integer",
          "format": "int32"
        },
        "systemId": {
          "type": "integer",
          "format": "int32"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "stateId": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "id",
        "title",
        "description",
        "severity",
        "environmentId",
        "systemId",
        "userId",
        "stateId"
      ]
    },
...
  },
...
}